+++ packages/desktop/src/renderer/components/widgets/channels/BasicMessage.tsx
@@ -152,7 +152,6 @@ export interface BasicMessageProps {
     src: string
   }>
   onMathMessageRendered?: () => void
-  userLabel?: UserLabelType
   unregisteredUsernameModalHandleOpen: HandleOpenModalType
   duplicatedUsernameModalHandleOpen: HandleOpenModalType
 }
@@ -169,9 +168,14 @@ export const BasicMessageComponent: React.FC<BasicMessageProps & FileActionsProp
   cancelDownload,
   unregisteredUsernameModalHandleOpen,
   duplicatedUsernameModalHandleOpen,
-  userLabel = null,
 }) => {
   const messageDisplayData = messages[0]
+
+  const userLabel = messageDisplayData?.isDuplicated
+    ? UserLabelType.DUPLICATE
+    : !messageDisplayData?.isRegistered
+    ? UserLabelType.UNREGISTERED
+    : null
 
   const infoMessage = messageDisplayData.type === 3 // 3 stands for MessageType.Info