+++ packages/desktop/src/renderer/components/widgets/channels/NestedMessageContent.tsx
@@ -3,7 +3,7 @@ import { styled } from '@mui/material/styles'
 import theme from '../../../theme'
 import classNames from 'classnames'
 import { Grid } from '@mui/material'
-import { AUTODOWNLOAD_SIZE_LIMIT, DisplayableMessage, DownloadState, DownloadStatus } from '@quiet/state-manager'
+import { AUTODOWNLOAD_SIZE_LIMIT, DownloadState, DownloadStatus } from '@quiet/state-manager'
 
 import UploadedImage from '../../Channel/File/UploadedImage/UploadedImage'
 import FileComponent, { FileActionsProps } from '../../Channel/File/FileComponent/FileComponent'
@@ -11,6 +11,7 @@ import { displayMathRegex } from '../../../../utils/functions/splitByTex'
 import { TextMessageComponent } from './TextMessage'
 import { MathMessageComponent } from '../../MathMessage/MathMessageComponent'
 import { UseModalType } from '../../../containers/hooks'
+import { DisplayableMessage } from '@quiet/types'
 
 const PREFIX = 'NestedMessageContent'