+++ packages/desktop/src/renderer/sagas/notifications/notifications.saga.ts
@@ -37,7 +37,7 @@ export function* displayMessageNotificationSaga(
   const publicChannelsSelector = yield* select(publicChannels.selectors.publicChannels)
 
   const currentIdentity = yield* select(identity.selectors.currentIdentity)
-  const certificatesMapping = yield* select(users.selectors.certificatesMapping)
+  const allUsers = yield* select(users.selectors.allUsers)
 
   const lastConnectedTime = yield* select(connection.selectors.lastConnectedTime)
 
@@ -54,7 +54,7 @@ export function* displayMessageNotificationSaga(
     if (focused && message.channelId === currentChannelId) return
 
     // Do not display notifications for own messages
-    const sender = certificatesMapping[message.pubKey]?.username
+    const sender = allUsers[message.pubKey]?.username
     if (!sender || sender === currentIdentity?.nickname) return
 
     // Do not display notifications if turned off in configuration