+++ packages/state-manager/src/types.ts
@@ -16,6 +16,7 @@ import {
   type SendMessagePayload,
   type SocketActionTypes,
   type UploadFilePayload,
+  CommunityMetadata,
 } from '@quiet/types'
 
 type EmitEvent<Payload> = (payload: Payload) => void
@@ -39,6 +40,7 @@ export interface EmitEvents {
   [SocketActionTypes.LEAVE_COMMUNITY]: () => void
   [SocketActionTypes.CREATE_NETWORK]: EmitEvent<Community>
   [SocketActionTypes.SAVE_USER_CSR]: EmitEvent<SaveCSRPayload>
+  [SocketActionTypes.SEND_COMMUNITY_METADATA]: EmitEvent<CommunityMetadata>
 }
 
 export type Socket = IOSocket<DefaultEventsMap, EmitEvents>